New Module: favoritestoggle
favoritestoggle offers toggles for all modules in the favorites group to be quickly accessed within lighttable.
It offers to toggle single modules marked as favorites for all selected images  (enable if any selected image does not
have the module enabled, and  disable if all have it enabled). A ctrl+click does create a duplicate on which the module
is toggled.